==> notebooks/00_start_here.ipynb <==
# About this Project:
<hr>

This project was built with Visual Studio Code, however it can be opened using any notebook-compatible tool, including Jupyter, Databricks, SageMaker, etc.

**Libraries Used**:
- Pandas
- Sodapy
- Matplotlib
- Seaborn
- Statsmodels
- SciPy

<br>

**Run these notebooks**:
1. 00_start_here     : (this notebook) Optional
2. 01_data_ingestion : Run this notebook, ingests and saves data
3. 02_EDA            : Optional
4. 03_Processing     : Run this notebook, processes and saves data
5. 04_Analysis       : Run this notebook, performs analysis
