==> notebooks/01_data_ingestion.ipynb <==
# 01 Data Ingestion
This notebook will ingest and save the data to a local data folder. While our climate data is already in CSV format, we will need to use APIs to ingest and save the accident and speed limit data.

### Imports & Setup

# Imports
import pandas as pd
from sodapy import Socrata

# Specify that we want to get data from data.calgary.ca. For the datasets we want, no API key is required.
client = Socrata("data.calgary.ca", None)

### Weather Data
This data was manually downloaded from climatedata.ca

# Load & display weather data
weather = pd.read_csv("../data/climate-daily.csv")
weather.head()

### Speed Limit Data

# Get data from Socrata API endpoint
speeds_raw = client.get("2bwu-t32v", limit=3700)

# Convert to pandas DataFrame
speeds_df = pd.DataFrame.from_records(speeds_raw)

# Display preview of data
speeds_df.head()

# Save data to our data folder
speeds_df.to_csv('../data/speed_limits.csv')

### Traffic Accident Data

# Get data from Socrata API endpoint
traffic_raw = client.get("35ra-9556", limit=50000)

# Convert to pandas DataFrame
traffic_df = pd.DataFrame.from_records(traffic_raw)


# View first five rows of our data
traffic_df.head()

# Save our data as a CSV
traffic_df.to_csv("../data/traffic_accidents.csv")

Our data is now in the data/ folder and can be used in the later steps.
