==> notebooks/02_EDA.ipynb <==
# 02 Exploratory Data Analysis

In this notebook we will explore our dataset and understand what needs to be done. Any work that we need to do on our data to make it clean will be done in the next notebook (03_process_data)

# Import required library
import pandas as pd

### Load Data

# Load in data into dataframes
climate_df = pd.read_csv("../data/climate-daily.csv")
speed_df = pd.read_csv("../data/speed_limits.csv")
accident_df = pd.read_csv("../data/traffic_accidents.csv")

# Print out three dataframes and the shape (rows x columns)

print("Exploring Climate Data:")
display(climate_df.head())
print("Data shape:", climate_df.shape)
print("\n ----- \n")

print("Exploring Speed Limit Data:")
display(speed_df.head())
print(speed_df.shape)
print("\n ----- \n")

print("Exploring Traffic Accident Data:")
display(accident_df.head())
print(accident_df.shape)

### Visually Inspect Data Types
Print the data type for each column

print(climate_df.dtypes)

print(speed_df.dtypes)

print(accident_df.dtypes)

### Check for Nulls/Missing Data

climate_df.isnull().sum()

speed_df.isnull().sum()

accident_df.isnull().sum()

### Check for Duplicates

print("Climate data:", climate_df.duplicated().sum())
print("Speed data:", speed_df.duplicated().sum())
print("Accident data:", accident_df.duplicated().sum())

### Observations for Data Cleaning:

Overall, the data is very clean. This is likely because it comes from reputable sources that likely did their own data cleaning before publishing it online. However, for this project there is some additional cleaning and preparation that I will do to the data. Some of the observations that I will act on includes:
1. Different column naming standards. Some data is all lowercase will some is all uppercase
2. The climate data has a large number of null values in the columns for snowfall. This is because a day with zero snowfall will have a NaN instead of a 0.
3. Our speed limit data has a few null values that we need to either fill in or remove
4. Street names has different spellings, naming conventions, and abbreviations.
5. Deerfoot is missing from the dataset.
