==> notebooks/03_Processing.ipynb <==
# 03 Processing

### Setup

#### Imports

import pandas as pd
import numpy as np

#### Load Data

# Load in dataframes
climate_raw = pd.read_csv("../data/climate-daily.csv")
speeds_raw = pd.read_csv("../data/speed_limits.csv")
traffic_raw = pd.read_csv("../data/traffic_accidents.csv")

# Show first 5 records of each DF
display(climate_raw.head())
display(speeds_raw.head())
display(traffic_raw.head())

### Clean Data

For each of our three datasets, we will:
1. Change column data types or value formats
2. Filter columns to only work with what we need
3. Rename columns
4. Remove/fill-in NaN values

#### Traffic Data

# Change start_dt column to a datetime format, and parse out the time part of the date
traffic_raw["start_dt"] = pd.to_datetime(traffic_raw["start_dt"])
traffic_raw["date"] = traffic_raw["start_dt"].dt.date

# Filter out columns
traffic_raw = traffic_raw[['date', 'incident_info', 'quadrant']]
traffic_raw.columns

# We have no nulls that need to be removed
traffic_raw.isna().sum()

#### Speed Limit Data

# Filter out columns we don't need
speeds_raw = speeds_raw[['street_name', 'speed', 'created_dt', 'modified_attribute_dt']]

# Rename select columns
speeds_raw.rename(columns={
    'speed': 'speed_limit',
    'created_dt': 'created_date',
    'modified_attribute_dt': 'modified_date'
}, inplace=True)

# Remove null values, using street name as the primary column
speeds_raw.dropna(subset=['street_name'], inplace=True)
speeds_raw.isna().sum()

#### Climate Data

# Change local_date to datetime format
climate_raw["LOCAL_DATE"] = pd.to_datetime(climate_raw["LOCAL_DATE"])

climate_raw = climate_raw[['LOCAL_DATE', 'MEAN_TEMPERATURE', 'MIN_TEMPERATURE', 'MAX_TEMPERATURE', 'TOTAL_PRECIPITATION', 'SNOW_ON_GROUND']]

# Rename columns to be all lowercase, and match the other two datasets
climate_raw.rename(str.lower, axis='columns', inplace=True)
climate_raw.rename(columns={'local_date':'date'}, inplace=True)

# For our records where snow_on_ground is null, fill them in with 0. This is because there is no (zero) snow on the ground
climate_raw['snow_on_ground'] = climate_raw['snow_on_ground'].fillna(0)

climate_raw.isna().sum()

### Process Data

# Refresher on our dataframes and data
display(climate_raw.head())
display(speeds_raw.head())
display(traffic_raw.head())

##### Join Weather and Traffic Data

# Aggregate the data to do a count of unique rows for each date, and rename the column
traffic_agg = traffic_raw[['date', 'incident_info']].groupby("date").count().reset_index()

# Convert date column from string to datetime
traffic_agg["date"] = pd.to_datetime(traffic_agg["date"])

# Show sample data
traffic_agg.head()

# Join our traffic accident data with our weather data. 
# Since there are some dates in the climate data that aren't in the traffic data, so we need to drop those rows
traffic_temperature_agg = pd.merge(traffic_agg, climate_raw, on='date', how='left')
traffic_temperature_agg = traffic_temperature_agg.dropna()

# Display sample data
traffic_temperature_agg.head()

# Save our cleaned and refined data to a CSV to avoid having to repeat all of the above steps
traffic_temperature_agg.to_csv('../data/traffic_temperature_agg.csv')

##### Join Speed Limit and Traffic Data

# Make a copy of our raw traffic DF
traffic_roads_parsed = traffic_raw.copy()

# Parse out direction and quadrants, to only get the road name
# For example, Deerfoot North and Deetfoot South has the same speed limit

pattern = (
    r"^\s*"
    r"(?:(Westbound|Eastbound|Northbound|Southbound)\s+)?"
    r"(.+?)"
    r"(?:\s+(NW|NE|SW|SE))?"
    r"\s*$"
)

traffic_roads_parsed[["direction", "location", "quadrant"]] = (
    traffic_roads_parsed["incident_info"]
    .str.strip()
    .str.extract(pattern)
)


# Display first five records
traffic_roads_parsed.head()

# The descriptions contain the exact location, but we only want the general road that the accident occured on
# For example, "Deerfoot North before Stoney Trail" should just be "Deerfoot"
keywords = r"after|before|at|near|approaching|between|and|to"

traffic_roads_parsed["street_name"] = (
    traffic_roads_parsed["location"]
    .str.lower()
    .str.split(rf"\s+({keywords})\b", n=1)
    .str[0]
    .str.strip()
)


# We will use this code multiple times, so putting it in a function will make it more reusable

def remove_street_name(dataframe):
    """
    This function removes any words found in the list from street names.
    "Stoney Trail" will simply become "Stoney"
    Take a dataframe as aan input.
    """

    # Words to remove
    words_to_remove = ['drive', 'street', 'trail', 'avenue', 'boulevard', 'blvd', 'ramp', 'dr', 'gate', 'av', 'rd', 'pl', 'wy', 'cr', 'st', 'bv', 'hl', 'tr', 'hi w']

    # Create a regex pattern: \b = word boundary, | = or
    pattern = r'\b(?:' + '|'.join(words_to_remove) + r')\b'

    # Replace with empty string, case-insensitive
    dataframe['road_name_clean'] = dataframe['street_name'].str.replace(pattern, '', case=False, regex=True)

    # Optional: strip extra whitespace
    dataframe['road_name_clean'] = dataframe['road_name_clean'].str.strip()

    return dataframe


# Call our function to clean the columns
traffic_roads_parsed = remove_street_name(traffic_roads_parsed)

# Sort the values by modified_date and drop any duplicates where there are mutliple speed limits for the same street
speeds_filtered = speeds_raw.sort_values('modified_date', ascending=False) \
                        .drop_duplicates(subset='street_name', keep='first')

# Ensure that all street names are lowercase
speeds_filtered['street_name'] = speeds_filtered['street_name'].str.lower()

# Similar to above, I'm removing the direction of the street, because Deerfoot NE is the same as Deerfoot SE
pattern = (
    r"^\s*"
    r"(.+?)"
    r"(?:\s+(nw|ne|sw|se))?"
    r"\s*$"
)

# Strip directions from street names
speeds_filtered[["street_name", "direction"]] = (
    speeds_filtered["street_name"]
    .str.strip()
    .str.extract(pattern)
)


# Call our above function to remove the prefix/suffix from our speed limit data
speeds_filtered = remove_street_name(speeds_filtered)

# One of the major roads in Calgary is called Deerfoot, however, Deerfoot is legally a provincial road and not a municipal road.
# This causes issues because the speed limit is not in our dataset, but it is a major road where many accidents occur.
# To fix this, we will manually add it.
deerfoot_row = {'street_name': 'deerfoot', 'speed_limit': 100.0, 'created_date': None, 'modified_date': None, 'direction': None, 'road_name_clean': 'deerfoot'}
speeds_filtered.loc[len(speeds_filtered)] = deerfoot_row

# Filter out any rows that has a duplicate road name
speeds_filtered = speeds_filtered.sort_values(by='speed_limit', ascending=False).drop_duplicates(subset=['road_name_clean'], keep='first')

# Merge our traffic/accident dataset with the speed limit data
roads_with_speeds = pd.merge(traffic_roads_parsed, speeds_filtered, on='road_name_clean', how='left')

# Display first five rows
roads_with_speeds.head()

# Throughout our work, the date has been reverted back to object. We need to make it a date again
roads_with_speeds['date'] = pd.to_datetime(roads_with_speeds['date'])

# Using our above dataframe, find the number of records for each date and speed limit.
# For this analysis, it doesn't matter if an accident occured on Deerfoot or Stoney trail, because both roads have a speed limit of 100 KM/h
accidents_with_speeds_agg = roads_with_speeds.groupby(['date', 'speed_limit']).size().reset_index(name='count')

# Display first five rows
accidents_with_speeds_agg.head()

# Save our data to a CSV to avoid having to do all of the above steps if something goes wrong
accidents_with_speeds_agg.to_csv("../data/accidents_with_speeds_agg.csv")
